tools/requirements.txt
# Python dependencies for Bass Hole asset tools
#
# Install with: pip install -r requirements.txt

Pillow>=9.0.0    # Image processing
numpy>=1.20.0    # Array operations for dithering
